Split typemaps into its own file
